requirements.txt
Flask==2.1.2
Flask_SQLAlchemy==2.5.1
tensorflow==2.8.0
Werkzeug==2.1.2
pillow==8.3.2
#pickle
#h5py
# TODO: Complete requirements and setup docker.